+++ hr_course/README.rst
@@ -1,5 +1,5 @@
 =========
-HR Course
+Hr Course
 =========
 
 .. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -14,13 +14,13 @@ HR Course
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fhr-lightgray.png?logo=github
-    :target: https://github.com/OCA/hr/tree/14.0/hr_course
+    :target: https://github.com/OCA/hr/tree/13.0/hr_course
     :alt: OCA/hr
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/hr-14-0/hr-14-0-hr_course
+    :target: https://translation.odoo-community.org/projects/hr-13-0/hr-13-0-hr_course
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/116/14.0
+    :target: https://runbot.odoo-community.org/runbot/116/13.0
     :alt: Try me on Runbot
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
@@ -38,10 +38,12 @@ Usage
 
 To create a Course go to Employees -> Courses.
 
-Fill the information and click assign attendees to continue. Assign them and
-then start the course.
+To create a Training Schedule go to Employees -> Courses Schedule.
 
-One it has finished you must enter the results and finish the course.
+Fill the information and click assign attendees to continue. Assign them and
+then start the course session.
+
+One it has finished you must enter the results and finish the course session.
 
 Bug Tracker
 ===========
@@ -49,7 +51,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/hr/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/hr/issues/new?body=module:%20hr_course%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/hr/issues/new?body=module:%20hr_course%0Aversion:%2013.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -80,6 +82,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/hr <https://github.com/OCA/hr/tree/14.0/hr_course>`_ project on GitHub.
+This module is part of the `OCA/hr <https://github.com/OCA/hr/tree/13.0/hr_course>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
