+++ hr_course/i18n/hr_course.pot
