+++ hr_course/models/__init__.py
@@ -1,2 +1,3 @@
 from . import hr_employee
 from . import hr_course
+from . import hr_course_schedule
