+++ hr_course/readme/DESCRIPTION.rst
@@ -1,2 +1,2 @@
-This module allows your to manage employee's training courses and all its
+This module allows your to manage employee's training courses, course dates and all its
 validation process.
